move setting store types
